Refactor util structure (#414)
* Rename debugger to debugging_util

* Rename inspector to inspection_util

* Rename miscellaneous to miscellaneous_util

* Rename topology_manager to topology_util

* There is a strange util in pymodi

* Rename tutor to tutorial_util

* Rename usage_instructor to usage_util
