+++ modi/util/strange_util.py
